Move string_format functions into builtin_functions
These functions are ultimately all called from within
string_format_builtin_functions.
